+++ soloud/src/audio/wavstream.rs
@@ -8,8 +8,21 @@ pub struct WavStream {
 
 impl WavStream {
     /// Get the length of the WavStream object
-    pub fn length(&mut self) -> f64 {
-        assert!(!self._inner.is_null());
+    pub fn length(&self) -> f64 {
         unsafe { ffi::WavStream_getLength(self._inner) }
+    }
+
+    /// Load a file to memory
+    pub fn load_to_mem(&mut self, path: &std::path::Path) -> Result<(), SoloudError> {
+        unsafe {
+            let path = path.to_str().unwrap();
+            let path = std::ffi::CString::new(path)?;
+            let ret = ffi::WavStream_loadToMem(self._inner, path.as_ptr());
+            if ret != 0 {
+                Err(SoloudError::Internal(SoloudErrorKind::from_i32(ret)))
+            } else {
+                Ok(())
+            }
+        }
     }
 }