+++ soloud/src/filter/bass.rs
@@ -16,7 +16,6 @@ pub struct BassboostFilter {
 impl BassboostFilter {
     /// Set filter params
     pub fn set_params(&mut self, delay: f32) -> Result<(), SoloudError> {
-        assert!(!self._inner.is_null());
         unsafe {
             let ret = soloud_sys::soloud::BassboostFilter_setParams(self._inner, delay);
             if ret != 0 {