+++ soloud/src/filter/dc.rs
@@ -9,7 +9,6 @@ pub struct DCRemovalFilter {
 impl DCRemovalFilter {
     /// Set filter params
     pub fn set_params(&mut self) -> Result<(), SoloudError> {
-        assert!(!self._inner.is_null());
         unsafe {
             let ret = soloud_sys::soloud::DCRemovalFilter_setParams(self._inner);
             if ret != 0 {
@@ -22,7 +21,6 @@ impl DCRemovalFilter {
 
     /// Set filter params with extra args
     pub fn set_params_ex(&mut self, delay: f32) -> Result<(), SoloudError> {
-        assert!(!self._inner.is_null());
         unsafe {
             let ret = soloud_sys::soloud::DCRemovalFilter_setParamsEx(self._inner, delay);
             if ret != 0 {