+++ soloud/src/filter/lofi.rs
@@ -17,7 +17,6 @@ pub struct LofiFilter {
 impl LofiFilter {
     /// Set filter params
     pub fn set_params(&mut self, samplerate: f32, bit_depth: f32) -> Result<(), SoloudError> {
-        assert!(!self._inner.is_null());
         unsafe {
             let ret = soloud_sys::soloud::LofiFilter_setParams(self._inner, samplerate, bit_depth);
             if ret != 0 {